# contracts/api.yaml
openapi: "3.1.0"
info:
  title: EcoReport API
  version: "0.1.0"
  description: |
    Environmental compliance reporting API for Russian market.
    Forms: 2-TP Waste, 2-TP Air, 2-TP Water per Order #757 (26.12.2024).
servers:
  - url: /api
    description: Next.js API routes (relative)
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (HttpOnly, Secure, SameSite=Lax)
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - NOT_FOUND
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
        traceId:
          type: string
...
    WasteClassifier:
      type: object
      required: [id, code, name, hazardClass]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        hazardClass:
          type: string
          enum: ["I", "II", "III", "IV", "V"]
        parentCode:
          type: string
    PollutantCode:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        casNumber:
          type: string
        hazardClass:
          type: string
          enum: ["I", "II", "III", "IV", "V"]
        pdkMr:
          type: number
        pdkSs:
          type: number
        unit:
          type: string
...
security:
  - session: []
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      security: []
      summary: Register new user with organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, orgName]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                name:
                  type: string
                  maxLength: 255
                orgName:
                  type: string
                  maxLength: 255
                inn:
                  type: string
                  pattern: '^\d{10,12}$'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /reports/{id}/pdf:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: generatePdf
      tags: [reports]
      summary: Generate PDF in Rosprirodnadzor format
      responses:
        "200":
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
  /reference/waste:
    get:
      operationId: searchWaste
      tags: [reference]
      summary: Search FKKO waste classifiers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search by code or name (min 2 chars)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteClassifier"
  /reference/pollutants:
    get:
      operationId: searchPollutants
      tags: [reference]
      summary: Search pollutant codes
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search by code, name, or CAS number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PollutantCode"

  # --- Calendar / Dashboard ---
...